requirements.txt
Flask==0.10.1
Flask-SocketIO==0.6.0
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.10.4
argparse==1.2.1
eventlet==0.17.4
gevent==1.1b3
gevent-socketio==0.3.6
gevent-websocket==0.9.5
greenlet==0.4.7
itsdangerous==0.24
puzpy==0.2.1
python-engineio==0.5.1
python-socketio==0.4.0
six==1.9.0
wsgiref==0.1.2
